create item components, modify burger menu
