+++ app/components/Items/Cast/Cast.tsx
@@ -1,0 +1,7 @@
+export default function Cast() {
+  return (
+    <div className="flex flex-col items-center motion-preset-pop">
+      <span className="text-orange-400 font-['DynaPuff'] text-xl">Cast</span>
+    </div>
+  );
+}
