+++ app/components/Items/Game/Game.tsx
@@ -1,0 +1,7 @@
+export default function Game() {
+  return (
+    <div className="flex flex-col items-center motion-preset-pop">
+      <span className="text-orange-400 font-['DynaPuff'] text-xl">Game</span>
+    </div>
+  );
+}
