+++ app/page.tsx
@@ -9,41 +9,19 @@ import {
   DiscordLogo,
   HandPointing,
   InstagramLogo,
-  PaperPlaneTilt,
   TwitterLogo,
 } from "@phosphor-icons/react";
+import Hula from "./components/Items/Hula/Hula";
+import Game from "./components/Items/Game/Game";
+import Cast from "./components/Items/Cast/Cast";
+import Studio from "./components/Items/Studio/Studio";
 
 function PageContent() {
   const [menuOpen, setMenuOpen] = useState(false);
   const [desktopMenuOpen, setDesktopMenuOpen] = useState(false);
   const [isMobile, setIsMobile] = useState(false);
-  const [email, setEmail] = useState("");
-  const [error, setError] = useState("");
-  const [success, setSuccess] = useState(false);
+
   const searchParams = useSearchParams();
-
-  const submitEmail = () => {
-    const emailRegex = /^[^\s@]+@[^\s@]+\.[^\s@]+$/;
-    if (emailRegex.test(email) && email.length > 0) {
-      console.log("Valid email:", email);
-      setSuccess(true);
-    } else {
-      setError("Please enter a valid email address.");
-      const emailInput = document.getElementById("email-input");
-      if (emailInput) {
-        emailInput.classList.add("motion-preset-shake");
-        setTimeout(() => {
-          emailInput.classList.remove("motion-preset-shake");
-        }, 1000);
-      }
-    }
-  };
-
-  const handleEmailChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-    setEmail(e.target.value);
-    setSuccess(false);
-    setError("");
-  };
 
   useEffect(() => {
     setIsMobile(window.innerWidth < 768);
@@ -60,17 +38,6 @@ function PageContent() {
     };
   }, []);
 
-  useEffect(() => {
-    const handleKeyDown = (e: KeyboardEvent) => {
-      if (e.key === "Enter") {
-        submitEmail();
-      }
-    };
-
-    window.addEventListener("keydown", handleKeyDown);
-    return () => window.removeEventListener("keydown", handleKeyDown);
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [email]);
   return (
     <div
       id="outer-container"
@@ -136,40 +103,12 @@ function PageContent() {
               />
             </Link>
           </div>
-          <span className="text-orange-400 font-['DynaPuff']">
-            Talk with Hula!
-          </span>
-          <div
-            className="bg-[rgba(55,58,71,0.6)] h-8 w-64 rounded-md flex flex-row justify-between pl-2 pr-1 !opacity-100"
-            id="email-input"
-          >
-            <input
-              placeholder="Email Address"
-              className={`bg-transparent ${
-                error ? "text-red-400" : "text-[#b8b7ad]"
-              } border-none outline-none w-full max-w-[100%]`}
-              type="email"
-              value={email}
-              onChange={(e) => {
-                handleEmailChange(e);
-              }}
-            />
-            <div
-              onClick={submitEmail}
-              className="p-1 cursor-pointer flex justify-end"
-            >
-              <div className="h-full w-[1px] bg-[#db8776] mr-1" />
-              <PaperPlaneTilt color="#ED684D" weight="fill" size="24px" />
-            </div>
-          </div>
-
-          <span className="text-orange-400 font-['DynaPuff'] text-center max-w-[300px]">
-            We&apos;ll also send you updates as we continue to work on the game.
-          </span>
-          {error ? <span className="text-red-400">{error}</span> : null}
-          {success ? (
-            <span className="text-green-400">Thanks for signing up!</span>
+          {searchParams.get("item") === "hula" || !searchParams.get("item") ? (
+            <Hula />
           ) : null}
+          {searchParams.get("item") === "game" ? <Game /> : null}
+          {searchParams.get("item") === "cast" ? <Cast /> : null}
+          {searchParams.get("item") === "studio" ? <Studio /> : null}
         </div>
       </div>
     </div>
